[refactor] : images: String -> images: List<String>

+++ SotenShopServer/src/main/kotlin/com/soten/shop/domain/auth/LoginResponse.kt
@@ -4,5 +4,5 @@ data class LoginResponse(
     val token: String,
     val refreshToken: String,
     val userName: String,
-    val userId: Long
+    val userId: Int
 )
