+++ SotenShopServer/src/main/kotlin/com/soten/shop/domain/jpa/BaseEntity.kt
@@ -8,7 +8,7 @@ abstract class BaseEntity {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	open var id: Long? = null
+	open var id: Int? = null
 
 	open var createdAt: LocalDateTime? = null
 	open var updatedAt: LocalDateTime? = null
