+++ SotenShopServer/src/main/kotlin/com/soten/shop/domain/product/ProductRepository.kt
@@ -4,26 +4,26 @@ import org.springframework.data.domain.Pageable
 import org.springframework.data.jpa.repository.JpaRepository
 import java.util.*
 
-interface ProductRepository : JpaRepository<Product, Long> {
+interface ProductRepository : JpaRepository<Product, Int> {
 
-	override fun findById(id: Long?): Optional<Product>
+	override fun findById(id: Int?): Optional<Product>
 
 	fun findAllByOrderByIdDesc(pageable: Pageable): List<Product>
 
 	fun findByCategoryIdAndIdGreaterThanOrderByIdDesc(
-		categoryId: Int?, id: Long, pageable: Pageable
+		categoryId: Int?, id: Int, pageable: Pageable
 	): List<Product>
 
 	fun findByCategoryIdAndIdLessThanOrderByIdDesc(
-		categoryId: Int?, id: Long, pageable: Pageable
+		categoryId: Int?, id: Int, pageable: Pageable
 	): List<Product>
 
 	fun findByIdGreaterThanAndNameLikeOrderByIdDesc(
-		id: Long, keyword: String, pageable: Pageable
+		id: Int, keyword: String, pageable: Pageable
 	): List<Product>
 
 	fun findByIdLessThanAndNameLikeOrderByIdDesc(
-		id: Long, keyword: String, pageable: Pageable
+		id: Int, keyword: String, pageable: Pageable
 	): List<Product>
 
 	fun findAllByCategoryIdOrderByIdDesc(categoryId: Int): List<Product>
